+++ src/HttpClient/GenericHttpClient.php
@@ -2,11 +2,16 @@
 
 namespace Mtarld\ApiPlatformMsBundle\HttpClient;
 
+use Mtarld\ApiPlatformMsBundle\Event\RequestEvent;
 use Mtarld\ApiPlatformMsBundle\Microservice\Microservice;
 use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 use Symfony\Contracts\HttpClient\Exception\ExceptionInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 use Symfony\Contracts\HttpClient\ResponseInterface;
+
+// Help opcache.preload discover always-needed symbols
+class_exists(RequestEvent::class);
 
 /**
  * @final
@@ -23,6 +28,7 @@ class GenericHttpClient implements ReplaceableHttpClientInterface
 
     private $serializer;
     private $httpClient;
+    private $dispatcher;
 
     /**
      * @var iterable<AuthenticationHeaderProviderInterface>
@@ -32,11 +38,13 @@ class GenericHttpClient implements ReplaceableHttpClientInterface
     public function __construct(
         SerializerInterface $serializer,
         HttpClientInterface $httpClient,
-        iterable $authenticationHeaderProviders = []
+        iterable $authenticationHeaderProviders = [],
+        ?EventDispatcherInterface $dispatcher = null
     ) {
         $this->serializer = $serializer;
         $this->httpClient = $httpClient;
         $this->authenticationHeaderProviders = $authenticationHeaderProviders;
+        $this->dispatcher = $dispatcher;
     }
 
     /**
@@ -73,7 +81,13 @@ class GenericHttpClient implements ReplaceableHttpClientInterface
             }
         }
 
-        return $this->httpClient->request($method, $uri, $options);
+        $response = $this->httpClient->request($method, $uri, $options);
+
+        if (null !== $this->dispatcher) {
+            $this->dispatcher->dispatch(new RequestEvent($microservice, $method, $uri, $options));
+        }
+
+        return $response;
     }
 
     public function setWrappedHttpClient(HttpClientInterface $httpClient): void