+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+* Added dispatching of a `RequestEvent` once a request is sent
+
 ## [0.3.0] - 2021-01-19
 * Added authentication header providers
 * Added PHP8 support